// frontend/src/css/type/theme.type.ts
export interface IBgColors {
  main: string;
  section: string;
  gradient: string;
  gray: string;
  purple: string;
}
export interface IColors {
  [key: string]: string;
}

export interface IFonts {
  title: string;
  sectionTitle: string;
  subTitle: string;
  text: string;
}
export interface IBolds {
  title: number;
  sectionTitle: number;
  subTitle: number;
  text: number;
}
export interface ISpacing {
  sm: number;
  md: number;
  lg: number;
}
export interface IBorderRadius {
  sm: number;
  md: number;
  lg: number;
}